Improved a guard test (fixed the code) and generalized a fixture.
