+++ tests/unit/test_messages_printer.py
@@ -498,52 +498,6 @@ def test_logfile_used(log_filepath):
 # -- tests for message showing external API
 
 
-class RecordingSpinner(_Spinner):
-    """A Spinner that records what is sent to supervision."""
-
-    def __init__(self, *args, **kwargs):
-        super().__init__(*args, **kwargs)
-        self.supervised = []
-
-    def supervise(self, message):
-        self.supervised.append(message)
-
-
-class RecordingPrinter(_Printer):
-    """A Printer isolated from outputs.
-
-    Instead, it records all messages to print.
-    """
-
-    def __init__(self, *args, **kwargs):
-        super().__init__(*args, **kwargs)
-        self.written_lines = []
-        self.written_bars = []
-        self.logged = []
-        self.spinner = RecordingSpinner(self)
-
-    def _write_line(self, message, *, spintext=None):
-        """Overwrite the real one to avoid it and record the message and maybe the spintext."""
-        if spintext is not None:
-            self.written_lines.append((message, spintext))
-        else:
-            self.written_lines.append(message)
-
-    def _write_bar(self, message):
-        """Overwrite the real one to avoid it and record the message."""
-        self.written_bars.append(message)
-
-    def _log(self, message):
-        """Overwrite the real one to avoid it and record the message."""
-        self.logged.append(message)
-
-
-@pytest.fixture
-def recording_printer(tmp_path):
-    """Provide a recording printer."""
-    return RecordingPrinter(tmp_path / "test.log")
-
-
 def test_show_defaults_no_stream(recording_printer):
     """Write a message with all defaults (without a stream)."""
     before = datetime.now()
